requirements.txt
flask
groq
python-dotenv
pandas
flask-cors 